+++ cube_builder/tasks.py
@@ -125,6 +125,8 @@ def warp_merge(activity, force=False):
             record.traceback = ''
             record.status = 'SUCCESS'
             record.args = merge_args
+
+            activity['args'].update(merge_args)
         except BaseException as e:
             record.status = 'FAILURE'
             record.traceback = capture_traceback(e)
@@ -158,9 +160,17 @@ def prepare_blend(merges):
         m['date']: (m['args']['efficacy'], m['args']['cloudratio']) for m in merges if m['band'] == 'quality'
     }
 
+    def _is_not_stk(_merge):
+        """Control flag to generate cloud mask.
+
+        This function is a utility to dispatch the cloud mask generation only for STK data cubes.
+        """
+        return _merge['band'] == 'quality' and not _merge['collection_id'].endswith('STK')
+
     for _merge in merges:
+        # Skip quality generation for MEDIAN, AVG
         if _merge['band'] in activities and _merge['args']['date'] in activities[_merge['band']]['scenes'] or \
-                _merge['band'] == 'quality':
+                _is_not_stk(_merge):
             continue
 
         activity = activities.get(_merge['band'], dict(scenes=dict()))
@@ -184,6 +194,8 @@ def prepare_blend(merges):
         }
 
         activities[_merge['band']] = activity
+
+    # TODO: Generate Vegetation Index
 
     logging.warning('Scheduling blend....')