+++ cube_builder/__init__.py
@@ -8,12 +8,14 @@
 
 """Python Module for Cube Builder."""
 
+from json import JSONEncoder
+
 from bdc_catalog.ext import BDCCatalog
 from flask import Flask
-from json import JSONEncoder
 from werkzeug.exceptions import HTTPException, InternalServerError
 
 from . import celery, config
+from .version import __version__
 
 
 def create_app(config_name=None):