+++ CHANGES.rst
@@ -1,5 +1,5 @@
 ..
-    This file is part of Python Module for Cube Builder.
+    This file is part of Brazil Data Cube Builder.
     Copyright (C) 2019-2020 INPE.
 
     Cube Builder is free software; you can redistribute it and/or modify it